requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
Django==1.11.4
django-celery==3.2.1
django-jquery==3.1.0
django-markdown==0.8.4
django-markup==1.2
django-mptt==0.8.7
docutils==0.14
kombu==3.0.37
Markdown==2.3.1
MySQL-python==1.2.5
Pygments==2.2.0
pytz==2017.2
six==1.10.0
